[CAP-41] files of career moved to separate package
