+++ app/src/main/java/org/scesi/cappuchino/data/database/subject/SubjectDao.kt
@@ -1,0 +1,4 @@
+package org.scesi.cappuchino.data.database.subject
+
+interface SubjectDao {
+}
