+++ data/src/main/java/org/scesi/data/dataSources/SubjectLocalDataSource.kt
@@ -1,0 +1,4 @@
+package org.scesi.data.dataSources
+
+interface SubjectLocalDataSource {
+}
